Replaces `live-server` with `five-server`
This replaces `live-server` with `five-server`, which is a fork/rewrite
of `live-server` which is maintained.

One issue that I found was that `five-server` adds a toast message upon
every page reload which is ugly, so I took the liberty of hiding that
info message. Other than that, it appears to work correctly.

Before this commit, `aiur` passed the `liveserver` parameter into faucet
and faucet took care of activating the liveserver (prompting the user to
install `live-server` if they hadn't already). I've modified the `aiur`
script so that it no longer passes that parameter into faucet, but
instead activates our own server implementation using `five-server`